Move stuff around and change documentation comments to “US Customary”
