+++ ResearchSuite/ResearchSuite/RSDTableItem.swift
@@ -104,462 +104,3 @@ open class RSDInputFieldTableItem : RSDTableItem {
         super.init(rowIndex: rowIndex)
     }
 }
-
-/// `RSDChoiceTableItem` is used to represent a single row in a table where the user picks from a list of choices.
-open class RSDChoiceTableItem : RSDInputFieldTableItem {
-    
-    /// The choice for a single or multiple choice input field.
-    open private(set) var choice: RSDChoice
-    
-    /// The answer associated with this choice
-    open override var answer: Any? {
-        return selected ? choice.value : nil
-    }
-    
-    /// Whether or not the choice is currently selected.
-    public var selected: Bool = false
-    
-    /// Initialize a new RSDChoiceTableItem.
-    /// parameters:
-    ///     - rowIndex:      The index of this item relative to all rows in the section in which this item resides.
-    ///     - inputField:    The RSDInputField representing this tableItem.
-    ///     - uiHint:        The UI hint for this row of the table.
-    ///     - choice:        The choice for a single or multiple choice input field.
-    public init(rowIndex: Int, inputField: RSDInputField, uiHint: RSDFormUIHint, choice: RSDChoice) {
-        self.choice = choice
-        super.init(rowIndex: rowIndex, inputField: inputField, uiHint: uiHint)
-    }
-}
-
-/// `RSDTextInputTableItem` is used to represent a single row in a table that holds a text entry input field.
-/// Typically, this would be used to represent a single `RSDInputField` value, but it can also be used to represent a
-/// single component in a multiple-component field.
-open class RSDTextInputTableItem : RSDInputFieldTableItem {
-    
-    /// The text field options for this input.
-    open private(set) var textFieldOptions: RSDTextFieldOptions?
-    
-    /// The placeholder text for this input.
-    open private(set) var placeholderText: String?
-    
-    /// The formatter used for dislaying answers and converting text to a number or date.
-    open private(set) var formatter: Formatter?
-    
-    /// The picker data source for picking an answer using a custom keyboard.
-    open private(set) var pickerSource: RSDPickerDataSource?
-    
-    /// The answer type for this component of the answer result.
-    public let answerType: RSDAnswerResultType
-    
-    /// Initialize a new RSDInputFieldTableItem.
-    /// parameters:
-    ///     - rowIndex: The index of this item relative to all rows in the section in which this item resides.
-    ///     - inputField: The RSDInputField representing this tableItem.
-    ///     - uiHint: The UI hint for this row of the table.
-    ///     - answerType: The answer type.
-    ///     - textFieldOptions: The text field options.
-    ///     - formatter: The formatter used for dislaying answers and converting text to a number or date.
-    ///     - pickerSource: The picker data source for selecting answers.
-    public init(rowIndex: Int, inputField: RSDInputField, uiHint: RSDFormUIHint, answerType: RSDAnswerResultType = .string, textFieldOptions: RSDTextFieldOptions? = nil, formatter: Formatter? = nil, pickerSource: RSDPickerDataSource? = nil, placeholderText: String? = nil) {
-        self.answerType = answerType
-        self.formatter = formatter
-        self.pickerSource = pickerSource
-        self.placeholderText = placeholderText ?? inputField.placeholderText
-        
-        // Set the text field options
-        self.textFieldOptions = textFieldOptions ?? inputField.textFieldOptions ?? {
-            switch answerType.baseType {
-            case .decimal:
-                return RSDTextFieldOptionsObject(keyboardType: .decimalPad)
-            case .integer, .timeInterval:
-                return RSDTextFieldOptionsObject(keyboardType: .numberPad)
-            case .date, .string:
-                return RSDTextFieldOptionsObject(keyboardType: .default)
-            case .boolean, .data:
-                return nil
-            }
-        }()
-        
-        super.init(rowIndex: rowIndex, inputField: inputField, uiHint: uiHint)
-    }
-    
-    /// The answer for this input field or component of a multiple component input field.
-    open override var answer: Any? {
-        return _answer
-    }
-    private var _answer: Any?
-    
-    /// The text string to display as the answer.
-    open var answerText: String? {
-        return answerText(for:_answer)
-    }
-    
-    /// The text string to display for a given answer.
-    open func answerText(for answer: Any?) -> String? {
-        if let text = pickerSource?.textAnswer(from: answer) ?? formatter?.string(for: answer) {
-            return text
-        } else if let anyAnswer = answer {
-            return String(describing: anyAnswer)
-        }
-        return nil
-    }
-    
-    /// Set the new answer value. This will throw an error if the value isn't valid. Otherwise, it will
-    /// set the answer.
-    /// - parameter newValue: The new value for the answer.
-    /// - throws: `RSDInputFieldError` if the answer is invalid.
-    public final func setAnswer(_ newValue: Any?) throws {
-        _answer = try validatedAnswer(newValue)
-    }
-    
-    /// Convert the input answer into a validated answer of a supported type, or throw an error if it fails validation.
-    /// - parameter newValue: The new value for the answer.
-    /// - returns: The converted answer.
-    open func validatedAnswer(_ newValue: Any?) throws -> Any? {
-        guard let newAnswer = newValue, !(newAnswer is NSNull) else {
-            return nil
-        }
-        let answer = try convertAnswer(newAnswer)
-        
-        // Look for a range on the new value if it was converted from a text field
-        if let _ = newValue as? String, (answer != nil) {
-            switch answerType.baseType {
-                
-            case .date:
-                if let date = answer as? Date, let range = inputField.range as? RSDDateRange {
-                    if let minDate = range.minimumDate, date < minDate {
-                        let context = RSDInputFieldError.Context(identifier: inputField.identifier, value: answer, answerResult: answerType, debugDescription: "Value entered is outside allowed range.")
-                        throw RSDInputFieldError.lessThanMinimumDate(minDate, context)
-                    }
-                    if let maxDate = range.maximumDate, date > maxDate {
-                        let context = RSDInputFieldError.Context(identifier: inputField.identifier, value: answer, answerResult: answerType, debugDescription: "Value entered is outside allowed range.")
-                        throw RSDInputFieldError.greaterThanMaximumDate(maxDate, context)
-                    }
-                }
-                
-            case .string:
-                if let string = answer as? String {
-                    if let validator = self.textFieldOptions?.textValidator, let isValid = try? validator.isValid(string), !isValid {
-                        let debugDescription = self.textFieldOptions?.invalidMessage ?? "Invalid regex"
-                        let context = RSDInputFieldError.Context(identifier: inputField.identifier, value: answer, answerResult: answerType, debugDescription: debugDescription)
-                        throw RSDInputFieldError.invalidRegex(self.textFieldOptions?.invalidMessage, context)
-                    }
-                    else if let maxLen = self.textFieldOptions?.maximumLength, maxLen > 0, string.count > maxLen {
-                        let context = RSDInputFieldError.Context(identifier: inputField.identifier, value: answer, answerResult: answerType, debugDescription: "Exceeds max length of \(maxLen)")
-                        throw RSDInputFieldError.exceedsMaxLength(maxLen, context)
-                    }
-                }
-                
-            default:
-                break
-            }
-        }
-        
-        return answer
-    }
-    
-    /// Convert the input answer into a validated answer of a supported type, or throw an error if it fails validation.
-    /// - parameter newValue: The new value for the answer.
-    /// - returns: The converted answer.
-    open func convertAnswer(_ newValue: Any) throws -> Any? {
-        var answer = newValue
-        
-        // First check if this is an array and if so, if it needs to have the first value pulled from it.
-        if let array = answer as? [Any] {
-            if answerType.sequenceType == .array {
-                return array
-            } else if array.count == 0 {
-                return nil
-            } else if array.count == 1 {
-                answer = array.first!
-            } else {
-                let context = RSDInputFieldError.Context(identifier: inputField.identifier, value: answer, answerResult: answerType, debugDescription: "Array Type \(answer) is not supported for \(inputField.identifier)")
-                throw RSDInputFieldError.invalidType(context)
-            }
-        }
-        
-        if answerType.baseType == .string {
-            return (answer as? String) ?? formatter?.string(for: answer) ?? "\(answer)"
-        }
-        else if let string = answer as? String {
-            if let formatter = self.formatter {
-                var obj: AnyObject?
-                var err: NSString?
-                formatter.getObjectValue(&obj, for: string, errorDescription: &err)
-                if err != nil {
-                    let context = RSDInputFieldError.Context(identifier: inputField.identifier, value: answer, answerResult: answerType, debugDescription: (err as String!))
-                    throw RSDInputFieldError.invalidFormatter(formatter, context)
-                } else {
-                    return obj
-                }
-            } else if answerType.baseType == .boolean {
-                return NSNumber(value: (string as NSString).boolValue)
-            } else if answerType.baseType == .integer {
-                return NSNumber(value: (string as NSString).integerValue)
-            } else if answerType.baseType == .decimal || answerType.baseType == .timeInterval {
-                return NSNumber(value: (string as NSString).doubleValue)
-            } else {
-                let context = RSDInputFieldError.Context(identifier: inputField.identifier, value: answer, answerResult: answerType, debugDescription: "String Type \(answer) is not supported for \(inputField.identifier)")
-                throw RSDInputFieldError.invalidType(context)
-            }
-        }
-        else if let date = answer as? Date {
-            if answerType.baseType == .date {
-                return date
-            } else {
-                let context = RSDInputFieldError.Context(identifier: inputField.identifier, value: answer, answerResult: answerType, debugDescription: "Date Type \(answer) is not supported for \(inputField.identifier)")
-                throw RSDInputFieldError.invalidType(context)
-            }
-        }
-        else if let num = (answer as? NSNumber) ?? (answer as? RSDJSONNumber)?.jsonNumber()  {
-            switch answerType.baseType  {
-            case .boolean:
-                return num.boolValue
-            case .integer, .decimal, .timeInterval:
-                return num
-            default:
-                let context = RSDInputFieldError.Context(identifier: inputField.identifier, value: answer, answerResult: answerType, debugDescription: "Number Type \(answer) is not supported for \(inputField.identifier)")
-                throw RSDInputFieldError.invalidType(context)
-            }
-        } else {
-            let context = RSDInputFieldError.Context(identifier: inputField.identifier, value: answer, answerResult: answerType, debugDescription: "\(answer) is not supported for \(inputField.identifier)")
-            throw RSDInputFieldError.invalidType(context)
-        }
-    }
-}
-
-/// An table item for entering a number value.
-final class RSDNumberInputTableItem : RSDTextInputTableItem {
-    
-    private var numberRange: RSDNumberRange?
-    
-    /// Initialize a new RSDInputFieldTableItem.
-    /// parameters:
-    ///     - rowIndex:     The index of this item relative to all rows in the section in which this item resides.
-    ///     - inputField:   The RSDInputField representing this tableItem.
-    ///     - uiHint:       The UI hint for this row of the table.
-    public init(rowIndex: Int, inputField: RSDInputField, uiHint: RSDFormUIHint) {
-        
-        var pickerSource: RSDPickerDataSource? = inputField as? RSDPickerDataSource
-        var formatter: Formatter? = (inputField.range as? RSDRangeWithFormatter)?.formatter
-        var range: RSDNumberRange? = (inputField.range as? RSDNumberRange)
-        
-        if inputField.dataType.baseType == .year, let dateRange = inputField.range as? RSDDateRange {
-            let calendar = Calendar(identifier: .gregorian)
-            let min: Int? = (dateRange.minimumDate != nil) ? calendar.component(.year, from: dateRange.minimumDate!) : nil
-            let max: Int? = (dateRange.maximumDate != nil) ? calendar.component(.year, from: dateRange.maximumDate!) : nil
-            if min != nil || max != nil {
-                range = RSDNumberRangeObject(minimumInt: min, maximumInt: max)
-            }
-        }
-        
-        let baseType: RSDAnswerResultType.BaseType = (inputField.dataType.baseType == .decimal) ? .decimal : .integer
-        let digits = (baseType == .decimal) ? 3 : 0
-        let numberFormatter = (formatter as? NumberFormatter) ?? NumberFormatter.defaultNumberFormatter(with: digits)
-        if inputField.dataType.baseType == .year {
-            numberFormatter.groupingSeparator = ""
-        }
-        formatter = formatter ?? numberFormatter
-        
-        if pickerSource == nil, let range = range, let min = range.minimumValue, let max = range.maximumValue {
-            pickerSource = RSDNumberPickerDataSourceObject(minimum: min, maximum: max, stepInterval: range.stepInterval, numberFormatter: numberFormatter)
-        }
-        
-        let answerType = RSDAnswerResultType(baseType: baseType, sequenceType: nil, dateFormat: nil, unit: range?.unit, sequenceSeparator: nil)
-        
-        self.numberRange = range
-        
-        super.init(rowIndex: rowIndex, inputField: inputField, uiHint: uiHint, answerType: answerType, textFieldOptions: nil, formatter: formatter, pickerSource: pickerSource)
-    }
-    
-    /// Convert the input answer into a validated answer of a supported type, or throw an error if it fails validation.
-    /// - parameter newValue: The new value for the answer.
-    /// - returns: The converted answer.
-    override func validatedAnswer(_ newValue: Any?) throws -> Any? {
-        guard let answer = try super.validatedAnswer(newValue) else {
-            return nil
-        }
-        
-        // Look for a range on the new value if it was converted from a text field
-        if let _ = newValue as? String {
-            switch answerType.baseType {
-            case .integer, .decimal, .timeInterval:
-                if let number = (answer as? NSNumber) ?? (answer as? RSDJSONNumber)?.jsonNumber(), let range = numberRange {
-                    let decimal = number.decimalValue
-                    if let min = range.minimumValue, decimal < min {
-                        let context = RSDInputFieldError.Context(identifier: inputField.identifier, value: answer, answerResult: answerType, debugDescription: "Value entered is outside allowed range.")
-                        throw RSDInputFieldError.lessThanMinimumValue(min, context)
-                    }
-                    if let max = range.maximumValue, decimal > max {
-                        let context = RSDInputFieldError.Context(identifier: inputField.identifier, value: answer, answerResult: answerType, debugDescription: "Value entered is outside allowed range.")
-                        throw RSDInputFieldError.greaterThanMaximumValue(max, context)
-                    }
-                }
-                
-            default:
-                break
-            }
-        }
-        
-        return answer
-    }
-}
-
-/// A custom implementation for inputing a person's mass (weight) that can set up a picker for
-/// setting the weight of an infant. Typically, US English user's know what their newborn baby's
-/// weight is in lb and oz. Additionally, the placeholder text for an adult or child measurement
-/// range will be either "pounds" or "kilograms", depending upon the user's locale.
-final class RSDMassInputTableItem : RSDTextInputTableItem {
-    
-    /// The base unit is the unit of mass that the measurement should be converted to in order
-    /// to save the result. Because the `Measurement` class has a generic `UnitType`, it cannot
-    /// be easily converted to a `Codable` object so the results are stored as `.decimal` base
-    /// type with the unit symbol also stored on the result.
-    public let baseUnit: UnitMass
-    
-    /// Default initializer.
-    /// - parameters:
-    ///     - rowIndex:         The index of this item relative to all rows in the section in which this item resides.
-    ///     - inputField:       The RSDInputField representing this tableItem.
-    ///     - uiHint:           The UI hint for this row of the table.
-    ///     - measurementSize:  The measurement range for the input field.
-    init(rowIndex: Int, inputField: RSDInputField, uiHint: RSDFormUIHint, measurementSize: RSDFormDataType.MeasurementRange) {
-        
-        // initial unit
-        var unit = (inputField.range as? RSDNumberRange)?.unit
-        
-        // Setup the formatter.
-        var formatter: Formatter? = (inputField.range as? RSDRangeWithFormatter)?.formatter
-        var placeholderText: String?
-        if (formatter == nil) {
-            let massFormatter = RSDMassFormatter(forInfantUse: (measurementSize == .infant), unitSymbol: unit)
-            formatter = massFormatter
-        }
-        
-        let massUnit: UnitMass
-        if let massFormatter = formatter as? RSDMassFormatter {
-            // When converting from the value entered by the participant, then the
-            // locale is used to determine the preferred units.
-            massFormatter.unitStyle = .long
-            if Locale.current.usesMetricSystem {
-                placeholderText = massFormatter.unitString(fromValue: 60, unit: .kilogram)
-            } else {
-                placeholderText = massFormatter.unitString(fromValue: 140, unit: .pound)
-            }
-            massFormatter.unitStyle = .medium
-            
-            massUnit = massFormatter.toStringUnit
-        }
-        else {
-            massUnit = ((unit != nil) ? UnitMass(fromSymbol: unit!) : nil) ?? .kilograms
-        }
-        
-        self.baseUnit = massUnit
-        unit = massUnit.symbol
-        
-        // If the measurement is for an infant and the local is US
-        // then use the infant mass picker.
-        var pickerSource: RSDPickerDataSource? = (inputField as? RSDPickerDataSource)
-        if pickerSource == nil, measurementSize == .infant, !Locale.current.usesMetricSystem {
-            pickerSource = RSDUSInfantMassPickerDataSource(formatter: formatter as? RSDMassFormatter)
-        }
-        
-        // Switch the hint type to a supported type most appropriate to the units.
-        var hint: RSDFormUIHint = uiHint
-        if uiHint != .popover {
-            hint = (pickerSource == nil) ? .textfield : .picker
-        }
-
-        let answerType = RSDAnswerResultType(baseType: .decimal, sequenceType: nil, dateFormat: nil, unit: unit, sequenceSeparator: nil)
-        
-        super.init(rowIndex: rowIndex, inputField: inputField, uiHint: hint, answerType: answerType, textFieldOptions: nil, formatter: formatter, pickerSource: pickerSource, placeholderText: placeholderText)
-    }
-    
-    /// Override the `convertAnswer()` function to convert the `Measurement` returned
-    /// by the formatter into a decimal value in the `baseUnit`.
-    override func convertAnswer(_ newValue: Any) throws -> Any? {
-        let answer = try super.convertAnswer(newValue)
-        guard let measurement = answer as? Measurement<UnitMass> else { return answer }
-        return measurement.converted(to: baseUnit).value
-    }
-}
-
-/// A custom implementation for inputing a person's height that can set up a picker for
-/// setting the height of an adult. Typically, US English user's will report an adult height
-/// using ft and in, whereas a child or infant will be reported in inches. The placeholder
-/// text for child or infant measurement range will be either "inches" or "centimeters",
-/// depending upon the user's locale.
-final class RSDHeightInputTableItem : RSDTextInputTableItem {
-    
-    /// The base unit is the unit of mass that the measurement should be converted to in order
-    /// to save the result. Because the `Measurement` class has a generic `UnitType`, it cannot
-    /// be easily converted to a `Codable` object so the results are stored as `.decimal` base
-    /// type with the unit symbol also stored on the result.
-    public let baseUnit: UnitLength
-    
-    /// Default initializer.
-    /// - parameters:
-    ///     - rowIndex:         The index of this item relative to all rows in the section in which this item resides.
-    ///     - inputField:       The RSDInputField representing this tableItem.
-    ///     - uiHint:           The UI hint for this row of the table.
-    ///     - measurementSize:  The measurement range for the input field.
-    init(rowIndex: Int, inputField: RSDInputField, uiHint: RSDFormUIHint, measurementSize: RSDFormDataType.MeasurementRange) {
-        
-        // initial unit
-        var unit = (inputField.range as? RSDNumberRange)?.unit
-        
-        // Setup the formatter.
-        var formatter: Formatter? = (inputField.range as? RSDRangeWithFormatter)?.formatter
-        var placeholderText: String?
-        if (formatter == nil) {
-            let lengthFormatter = RSDLengthFormatter(forChildUse: (measurementSize != .adult), unitSymbol: unit)
-            formatter = lengthFormatter
-        }
-        
-        let lengthUnit: UnitLength
-        if let lengthFormatter = formatter as? RSDLengthFormatter {
-            // When converting from the value entered by the participant, then the
-            // locale is used to determine the preferred units.
-            lengthFormatter.unitStyle = .long
-            if Locale.current.usesMetricSystem {
-                placeholderText = lengthFormatter.unitString(fromValue: 250, unit: .centimeter)
-            } else {
-                placeholderText = lengthFormatter.unitString(fromValue: 60, unit: .inch)
-            }
-            lengthFormatter.unitStyle = .short
-            
-            lengthUnit = lengthFormatter.toStringUnit
-        }
-        else {
-            lengthUnit = ((unit != nil) ? UnitLength(fromSymbol: unit!) : nil) ?? .centimeters
-        }
-        
-        self.baseUnit = lengthUnit
-        unit = lengthUnit.symbol
-        
-        // If the measurement size is for an adult and the locale is US
-        // then use a picker with feet/inches.
-        var pickerSource: RSDPickerDataSource? = (inputField as? RSDPickerDataSource)
-        if pickerSource == nil, measurementSize == .adult, !Locale.current.usesMetricSystem {
-            pickerSource = RSDUSHeightPickerDataSource(formatter: formatter as? RSDLengthFormatter)
-        }
-        
-        // Switch the hint type to a supported type.
-        var hint: RSDFormUIHint = uiHint
-        if uiHint != .popover {
-            hint = (pickerSource == nil) ? .textfield : .picker
-        }
-        
-        let answerType = RSDAnswerResultType(baseType: .decimal, sequenceType: nil, dateFormat: nil, unit: unit, sequenceSeparator: nil)
-        
-        super.init(rowIndex: rowIndex, inputField: inputField, uiHint: hint, answerType: answerType, textFieldOptions: nil, formatter: formatter, pickerSource: pickerSource, placeholderText: placeholderText)
-    }
-    
-    /// Override the `convertAnswer()` function to convert the `Measurement` returned
-    /// by the formatter into a decimal value in the `baseUnit`.
-    override func convertAnswer(_ newValue: Any) throws -> Any? {
-        let answer = try super.convertAnswer(newValue)
-        guard let measurement = answer as? Measurement<UnitLength> else { return answer }
-        return measurement.converted(to: baseUnit).value
-    }
-}
